+++ include/sentry.h
@@ -2300,6 +2300,19 @@ SENTRY_API void sentry_scope_set_context(
     sentry_scope_t *scope, const char *key, sentry_value_t value);
 SENTRY_API void sentry_scope_set_context_n(sentry_scope_t *scope,
     const char *key, size_t key_len, sentry_value_t value);
+/**
+ * Updates a context object by merging the provided values into it.
+ * Keys in `value` take precedence over keys in the context, updating the
+ * existing values. Keys in the existing context that are not present in `value`
+ * are preserved. If the context object does not exist yet, it will be created.
+ */
+SENTRY_API void sentry_update_context(const char *key, sentry_value_t value);
+SENTRY_API void sentry_update_context_n(
+    const char *key, size_t key_len, sentry_value_t value);
+SENTRY_API void sentry_scope_update_context(
+    sentry_scope_t *scope, const char *key, sentry_value_t value);
+SENTRY_API void sentry_scope_update_context_n(sentry_scope_t *scope,
+    const char *key, size_t key_len, sentry_value_t value);
 
 /**
  * Removes the context object with the specified key.